requirements.txt
alembic==1.1.0
amqp==2.5.1
aniso8601==7.0.0
atomicwrites==1.3.0
attrs==19.1.0
billiard==3.6.1.0
celery==4.3.0
Click==7.0
factory-boy==2.12.0
Faker==2.0.1
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
gitpython==3.0.2
gunicorn==19.9.0
importlib-metadata==0.20
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.6.4
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
psycopg2-binary==2.8.3
py==1.8.0
pyparsing==2.4.2
pytest==5.1.2
python-dateutil==2.8.0
python-editor==1.0.4
pytest-cov==2.7.1
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.8
SQLAlchemy-Utils==0.34.2
text-unidecode==1.2
vine==1.3.0
wcwidth==0.1.7
Werkzeug==0.15.6
zipp==0.6.0
